--- a/src/core.js
+++ b/src/core.js
@@ -2120,6 +2120,8 @@ define([
 
     fn.contributeToHeadXML = function (xmlWriter, form) {}; 
 
+    fn.addAutocomplete = function (input, form, options) {};
+
     fn.initWidget = function (widget) {};
 
     fn.destroy = function () {};