--- a/src/uploader.js
+++ b/src/uploader.js
@@ -371,6 +371,7 @@ define([
                 }
                 delete control.value;
             });
+            this.__callOld();
         }
     });
 });