--- a/src/writer.js
+++ b/src/writer.js
@@ -68,9 +68,6 @@ define([
         }
 
         uuid = form.formUuid; //gets set at parse time/by UI
-        if(!uuid) {
-            uuid = "http://openrosa.org/formdesigner/" + util.generate_xmlns_uuid();
-        }
 
         xmlWriter.writeAttributeString("xmlns:jrm",jrm);
         xmlWriter.writeAttributeString("xmlns", uuid);