Turns out cool features slow down tests

Before: `./test  103.82s user 1.18s system 96% cpu 1:48.64 total`
After: `./test  60.94s user 1.32s system 95% cpu 1:05.35 total`

Removes atwho from vellum tests. Atwho is currently tested by creating it's own input and has a mock form.

Depends on https://github.com/dimagi/commcare-hq/pull/9859

@millerdev 

cc: @sravfeyn