@emord 2016-01-12
Elapsed times from travis are down from 5:07 to 4:13