@millerdev 2016-01-13
Let me know what you think about the comments on the atwho plugin cache. I think this is good to merge if you agree that there is no leak.